BrowserAgentProject/requirements.txt
langchain-google-genai
langchain-core
langgraph
fastmcp
python-dotenv
playwright
